requirements.txt
streamlit
yt-dlp
openai-whisper
pydub
requests
deep-translator
pandas


